src/r/requirements.txt
# For llama.cpp core functionality
llama-cpp-python
psutil
pynvml

# For the model download script
huggingface_hub

# For the web server
fastapi
uvicorn[standard]